refac: move game repository to application level using another interface (abstraction)
